Restructured rating package to proper format. Added cronjob to main
