+++ cmd/main.go
@@ -3,10 +3,10 @@ package main
 import (
 	"fmt"
 	"github.com/NTNU-sondrbaa-2019/CLOUD-O1/pkg/CO1Cache"
-	"github.com/NTNU-sondrbaa-2019/CLOUD-PROJECT/internal"
-	"net/http"
-
+	"github.com/NTNU-sondrbaa-2019/CLOUD-PROJECT/internal/rating"
 )
+
+import "github.com/robfig/cron/v3"
 
 func main() {
 
@@ -16,19 +16,30 @@ func main() {
 	}
 
 	test := Test {
-		"This is a testa JSON",
+		"This is a test JSON",
 		"Sondre Benjamin Aasen",
 	}
 
-	http.HandleFunc("/get/fake", internal.FakeTeamMembers)
-	http.HandleFunc("/get/member", internal.GetTeamMembers)
+
+
 	CO1Cache.Initialize()
 	CO1Cache.WriteJSON("test", test)
 
 	fmt.Println("Hello World!")
 
-	http.ListenAndServe(":8080", nil)
+	// Uncomment to run the lichess stuff.
 
+	c := cron.New()
+	teamIdKey := "storbukk-sjakklubb"
+	_, err := c.AddFunc("0 2 * * *", func() {
+		rating.GetTeamElo(teamIdKey)
+	})
+
+	if err != nil {
+		panic(err)
+	}
+
+	c.Start()
 }
