+++ go.mod
@@ -2,4 +2,7 @@ module github.com/NTNU-sondrbaa-2019/CLOUD-PROJECT
 
 go 1.13
 
-require github.com/NTNU-sondrbaa-2019/CLOUD-O1 v0.0.0-20191020204558-b35729f3a3a7
+require (
+	github.com/NTNU-sondrbaa-2019/CLOUD-O1 v0.0.0-20191020204558-b35729f3a3a7
+	github.com/robfig/cron/v3 v3.0.0 // indirect
+)
