+++ internal/rating/fakeTeamMembers.go
@@ -1,0 +1,47 @@
+package rating
+
+/*
+func FakeTeamMembers() []TeamMember{
+	var Teamids[] TeamMember
+	Teamids[0].ID = "iwindj"
+	Teamids[1].ID = "JoakimPB"
+	Teamids[2].ID = "BeagluZ"
+	Teamids[3].ID = "NorTroll"
+	Teamids[4].ID = "Atman96"
+	Teamids[5].ID = "HermanDyrkorn"
+	Teamids[6].ID = "Y3SH1"
+	Teamids[7].ID = "TyrotoxismB"
+	Teamids[8].ID = "noDiva"
+	Teamids[9].ID = "StorbukkSK"
+	Teamids[10].ID = "Hyge"
+	Teamids[11].ID = "andreas_nl"
+
+	Teamids[0].Username = "iwindj"
+	Teamids[1].Username = "JoakimPB"
+	Teamids[2].Username = "BeagluZ"
+	Teamids[3].Username = "NorTroll"
+	Teamids[4].Username = "Atman96"
+	Teamids[5].Username = "HermanDyrkorn"
+	Teamids[6].Username = "Y3SH1"
+	Teamids[7].Username = "TyrotoxismB"
+	Teamids[8].Username = "noDiva"
+	Teamids[9].Username = "StorbukkSK"
+	Teamids[10].Username = "Hyge"
+	Teamids[11].Username = "andreas_nl"
+
+	Teamids[0].InternalCreatedAt = 1572566400
+	Teamids[1].InternalCreatedAt = 1572566400
+	Teamids[2].InternalCreatedAt = 1572566400
+	Teamids[3].InternalCreatedAt = 1572566400
+	Teamids[4].InternalCreatedAt = 1572566400
+	Teamids[5].InternalCreatedAt = 1572566400
+	Teamids[6].InternalCreatedAt = 1572566400
+	Teamids[7].InternalCreatedAt = 1572566400
+	Teamids[8].InternalCreatedAt = 1572566400
+	Teamids[9].InternalCreatedAt = 1572566400
+	Teamids[10].InternalCreatedAt = 1572566400
+	Teamids[11].InternalCreatedAt = 1572566400
+
+	return Teamids
+}
+*/
