+++ internal/rating/structs.go
@@ -1,0 +1,31 @@
+package rating
+
+type TeamMember struct {
+	ID string `json:"id"`
+	Username string `json:"username"`
+	InternalCreatedAt int  `json:"internalCreatedAt"`
+	Elo float64 `json:"elo"`
+}
+
+//TeamMember - Struct for getting essential teammembers information
+type Game struct {
+	ID        string `json:"id"`
+	CreatedAt int  `json:"createdAt"`
+	Players   struct {
+		White struct {
+			User struct {
+				ID   string `json:"id"`
+				Name string `json:"name"`
+			} `json:"user"`
+			Rating int `json:"rating"`
+		} `json:"white"`
+		Black struct {
+			User struct {
+				ID   string `json:"id"`
+				Name string `json:"name"`
+			} `json:"user"`
+			Rating int `json:"rating"`
+		} `json:"black"`
+	} `json:"players"`
+	Winner string `json:"winner"`
+}
